feat(CU-#86bxx9f5d): add state for calendar component
