+++ .eslintrc.js
@@ -48,7 +48,7 @@ module.exports = {
         "import/extensions": 0,
         "import/prefer-default-export": 0,
         "no-unused-vars": "off",
-        "@typescript-eslint/no-unused-vars": "error",
+        // "@typescript-eslint/no-unused-vars": "error",
         "@typescript-eslint/no-explicit-any": "error",
         "react-hooks/exhaustive-deps": "error",
         "prettier/prettier": ["error", { endOfLine: "auto" }],
@@ -60,5 +60,6 @@ module.exports = {
         "@typescript-eslint/naming-convention": "off",
         "no-restricted-exports": "off",
         "react/require-default-props": "off",
+        "@typescript-eslint/no-unused-vars": "off",
     },
 };
