+++ package.json
@@ -81,6 +81,7 @@
     "rollup-plugin-copy": "^3.5.0",
     "rollup-plugin-dts": "^6.1.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-tsconfig-paths": "^1.5.2",
     "rollup-plugin-visualizer": "^5.12.0",
     "storybook": "^8.0.0",
     "styled-components": "^6.1.8",
