+++ rollup.config.mjs
@@ -6,7 +6,8 @@ import terser from "@rollup/plugin-terser";
 import peerDepsExternal from "rollup-plugin-peer-deps-external";
 import alias from '@rollup/plugin-alias';
 import copy from 'rollup-plugin-copy';
-import svgr from '@svgr/rollup'
+import svgr from '@svgr/rollup';
+import tsConfigPaths from "rollup-plugin-tsconfig-paths"
 
 
 const isDev = process.env.NODE_ENV === "development";
@@ -25,6 +26,7 @@ export default [
       },
     ],
     plugins: [
+      tsConfigPaths(),
       peerDepsExternal(),
       resolve(),
       commonjs(),
@@ -41,23 +43,15 @@ export default [
       }) : null,
       alias({
         entries: [
-          { find: '@', replacement: './src/' },
-          { find: '@/components/DatePicker', replacement: './src/components/DatePicker/DatePicker' }
+          { find: '@', replacement: './src' },
+          { find: '@/components/DatePicker', replacement: './src/components/DatePicker/DatePicker' },
         ]
       }),
       copy({
-        /*targets: [
-          { src: 'src/index.html', dest: 'dist/public' },
-          { src: [
-            'src/assets/fonts/OpenSans-Bold.ttf',
-            'src/assets/fonts/OpenSans-Regular.ttf',
-            'src/assets/fonts/OpenSans-SemiBold.ttf',
-            ], dest: 'dist/public/fonts' },
-          { src: 'assets/images/!**!/!*', dest: 'dist/public/images' }
-        ]*/
+
         targets: [{ src: 'src/assets/*', dest: 'dist/public/assets' }]
       }),
-      svgr()
+      svgr({ exportType: 'named', jsxRuntime: 'automatic' }),
     ]
   },
 ];
