+++ src/constants/calendar.ts
@@ -1,0 +1,26 @@
+export const REVERSE_DAYS = ["Su", "Mo", "Tu", "We", "Th", "Fr", "Sa"];
+export const DEFAULT_DAYS = ["Mo", "Tu", "We", "Th", "Fr", "Sa", "Su"];
+
+export const MONTH_NAMES: Record<string, string> = {
+    "01": "January",
+    "1": "January",
+    "02": "February",
+    "2": "February",
+    "03": "March",
+    "3": "March",
+    "04": "April",
+    "4": "April",
+    "05": "May",
+    "5": "May",
+    "06": "June",
+    "6": "June",
+    "07": "July",
+    "7": "July",
+    "08": "August",
+    "8": "August",
+    "09": "September",
+    "9": "September",
+    "10": "October",
+    "11": "November",
+    "12": "December",
+};
